requirements.txt
python-binance
typer
loguru
python-dotenv